move cargo invocation to module
